sql: TestServer shim infrastructure
This change adds infrastructure for a TestServerShim which can be used to
interface with TestServer without depending on `server`.

The central trick is that `TestMain` can run from the `sql_main` package
which has access to `server` and can initialize the shim. The initialization
could also happen as part of an `init()` function in `sql_main`. Kudos to
@AndreiMatei and @tamird for coming up with this idea.

This will allow us to run tests from the `sql` package and access `sql`
internals, as shown by a (temporary) proof-of-concept test. The same will apply
to `dist_sql` stuff when it moves to a separate package.