sql: clean up uses of Statement
Throughout the executor we pass around a Statement by value, but it
escapes to heap in various places (we store a pointer to it in the
planner). This change switches to allocating it upfront and passing it
by pointer.

We also add some constructors and clean up the paths where a Statement
is created. Finally, we always generate the AnonymizedStr - currently,
we only populate it in some cases (executing prepared statements) but
we always generate it sooner or later (to update query stats).

Release note: None